Criação da migration. Inserção do connection no appsetings. Alteração do Program para funcionar com o SQL Server. Modificação do TotalCost para double.

+++ DevFreela.Application/InputModels/NewProjectInputModel.cs
@@ -6,6 +6,6 @@
         public string Description { get; set; }
         public int IdClient { get; set; }
         public int IdFreelancer { get; set; }
-        public decimal TotalCost { get; set; }
+        public double TotalCost { get; set; }
     }
 }
