+++ DevFreela.Core/Entities/Project.cs
@@ -4,7 +4,7 @@ namespace DevFreela.Core.Entities
 {
     public class Project : BaseEntity
     {
-        public Project(string title, string description, int idClient, int idFreelancer, decimal totalCost)
+        public Project(string title, string description, int idClient, int idFreelancer, double totalCost)
         {
             Title = title;
             Description = description;
@@ -23,7 +23,7 @@ namespace DevFreela.Core.Entities
         public User Client { get; private set; }
         public int IdFreelancer { get; private set; }
         public User Freelancer { get; private set; }
-        public decimal TotalCost { get; private set; }
+        public double TotalCost { get; private set; }
         public DateTime CreatedAt { get; private set; }
         public DateTime? StartedAt { get; private set; }
         public DateTime? FinisheAt { get; private set; }
@@ -58,7 +58,7 @@ namespace DevFreela.Core.Entities
         /// <summary>
         /// Vai ser retirado posteriormente, por causa do Entity Framework
         /// </summary>
-        public void Update(string title, string description, decimal totalCost)
+        public void Update(string title, string description, double totalCost)
         {
             Title = title;
             Description = description;
